+++ torch/onnx/_internal/diagnostics/infra/utils.py
@@ -1,9 +1,11 @@
 import inspect
 from typing import Any, Callable, Dict, Mapping, Tuple
 
+from torch.onnx._internal import _beartype
 from torch.onnx._internal.diagnostics.infra import _infra, formatter
 
 
+@_beartype.beartype
 def python_frame(frame: inspect.FrameInfo) -> _infra.StackFrame:
     """Returns a StackFrame for the given inspect.FrameInfo."""
     snippet = (
@@ -23,6 +25,7 @@ def python_frame(frame: inspect.FrameInfo) -> _infra.StackFrame:
     )
 
 
+@_beartype.beartype
 def python_call_stack(frames_to_skip: int = 0, frames_to_log: int = 16) -> _infra.Stack:
     """Returns the current Python call stack."""
     if frames_to_skip < 0:
@@ -42,6 +45,7 @@ def python_call_stack(frames_to_skip: int = 0, frames_to_log: int = 16) -> _infr
     return stack
 
 
+@_beartype.beartype
 def function_location(fn: Callable) -> _infra.Location:
     """Returns a Location for the given function."""
     source_lines, lineno = inspect.getsourcelines(fn)
@@ -54,6 +58,7 @@ def function_location(fn: Callable) -> _infra.Location:
     )
 
 
+@_beartype.beartype
 def function_state(
     fn: Callable, args: Tuple[Any, ...], kwargs: Dict[str, Any]
 ) -> Mapping[str, Any]:
